Allow reading replacements from config
