+++ src/Microsoft.TemplateEngine.Core.Contracts/IReplacementTokens.cs
@@ -4,6 +4,6 @@
     {
         string VariableName { get; }
 
-        string OriginalValue { get; }
+        ITokenConfig OriginalValue { get; }
     }
 }
