+++ src/Microsoft.TemplateEngine.Core/Util/ProcessorState.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
@@ -8,16 +8,38 @@ using Microsoft.TemplateEngine.Core.Matching;
 
 namespace Microsoft.TemplateEngine.Core.Util
 {
-    public class ProcessorState2 : IProcessorState
+    public class ProcessorState : IProcessorState
     {
-        private readonly int _flushThreshold;
         private readonly Stream _source;
         private readonly Stream _target;
         private readonly TrieEvaluator<OperationTerminal> _trie;
         private Encoding _encoding;
         private static readonly ConcurrentDictionary<IReadOnlyList<IOperationProvider>, Dictionary<Encoding, Trie<OperationTerminal>>> TrieLookup = new ConcurrentDictionary<IReadOnlyList<IOperationProvider>, Dictionary<Encoding, Trie<OperationTerminal>>>();
+        private readonly int _flushThreshold;
 
-        public ProcessorState2(Stream source, Stream target, int bufferSize, int flushThreshold, IEngineConfig config, IReadOnlyList<IOperationProvider> operationProviders)
+        public IEngineConfig Config { get; }
+
+        public byte[] CurrentBuffer { get; }
+
+        public int CurrentBufferLength { get; private set; }
+
+        public int CurrentBufferPosition { get; private set; }
+
+        public int CurrentSequenceNumber { get; private set; }
+
+        public IEncodingConfig EncodingConfig { get; private set; }
+
+        public Encoding Encoding
+        {
+            get { return _encoding; }
+            set
+            {
+                _encoding = value;
+                EncodingConfig = new EncodingConfig(Config, _encoding);
+            }
+        }
+
+        public ProcessorState(Stream source, Stream target, int bufferSize, int flushThreshold, IEngineConfig config, IReadOnlyList<IOperationProvider> operationProviders)
         {
             //Buffer has to be at least as large as the largest BOM we could expect
             if (bufferSize < 4)
@@ -30,7 +52,7 @@ namespace Microsoft.TemplateEngine.Core.Util
                 {
                     if (source.Length < bufferSize)
                     {
-                        bufferSize = (int) source.Length;
+                        bufferSize = (int)source.Length;
                     }
                 }
                 catch
@@ -69,7 +91,7 @@ namespace Microsoft.TemplateEngine.Core.Util
                         {
                             if (op.Tokens[j] != null)
                             {
-                                trie.AddPath(op.Tokens[j].Value, new OperationTerminal(op, j, op.Tokens[j].Value.Length, op.Tokens[j].Start, op.Tokens[j].End));
+                                trie.AddPath(op.Tokens[j].Value, new OperationTerminal(op, j, op.Tokens[j].Length, op.Tokens[j].Start, op.Tokens[j].End));
                             }
                         }
                     }
@@ -91,207 +113,183 @@ namespace Microsoft.TemplateEngine.Core.Util
             }
         }
 
-        public IEngineConfig Config { get; }
-
-        public byte[] CurrentBuffer { get; }
-
-        public int CurrentBufferLength { get; private set; }
-
-        public int CurrentBufferPosition { get; private set; }
-
-        public int CurrentSequenceNumber { get; private set; }
-
-        public Encoding Encoding
-        {
-            get { return _encoding; }
-            set
-            {
-                _encoding = value;
-                EncodingConfig = new EncodingConfig(Config, _encoding);
-            }
-        }
-
-        public IEncodingConfig EncodingConfig { get; private set; }
-
         public bool AdvanceBuffer(int bufferPosition)
         {
-            if (CurrentBufferLength == 0)
+            if(CurrentBufferLength == 0 || bufferPosition == 0)
             {
+                return false;
+            }
+
+            //The number of bytes away from the current buffer position being 
+            //  retargeted to the buffer head
+            int netMove = bufferPosition - CurrentBufferPosition;
+            //Since the CurrentSequenceNumber and CurrentBufferPosition are
+            //  different mappings over the same value, the same net move
+            //  applies to the current sequence number
+            CurrentSequenceNumber += netMove;
+            //Calculate the number of bytes at the end of the buffer that
+            //  should be preserved
+            int bytesToPreserveInBuffer = CurrentBufferLength - bufferPosition;
+
+            if (CurrentBufferLength < CurrentBuffer.Length && bytesToPreserveInBuffer == 0)
+            {
+                CurrentBufferLength = 0;
                 CurrentBufferPosition = 0;
                 return false;
             }
 
-            if(bufferPosition == 0)
+            //If we actually have to preserve any data, shift it to the start
+            if (bytesToPreserveInBuffer > 0)
             {
-                return false;
+                //Shift the relevant number of bytes back to the head of the buffer
+                Buffer.BlockCopy(CurrentBuffer, bufferPosition, CurrentBuffer, 0, bytesToPreserveInBuffer);
             }
 
-            //At this point we know that CurrentBufferPosition and CurrentSequenceNumber are related
-            //  and we know that CurrentBufferPosition will be set to the head of the buffer at the
-            //  end of this method, shifting off bufferPosition bytes. The number of bytes that then
-            //  would be advanced in the sequence is then the number of bytes the bufferPosition was
-            //  ahead of the CurrentBuffer position
-            CurrentSequenceNumber += bufferPosition - CurrentBufferPosition;
+            //Fill the remaining spaces in the buffer with new data, save how
+            //  many we've read for recalculating the new effective buffer size
+            int nRead = _source.Read(CurrentBuffer, bytesToPreserveInBuffer, CurrentBufferLength - bytesToPreserveInBuffer);
+            CurrentBufferLength = bytesToPreserveInBuffer + nRead;
 
-            int offset = 0;
-            if (bufferPosition != CurrentBufferLength)
-            {
-                offset = CurrentBufferLength - bufferPosition;
-                Array.Copy(CurrentBuffer, bufferPosition, CurrentBuffer, 0, offset);
-            }
-
-            int bytesRead = _source.Read(CurrentBuffer, offset, CurrentBuffer.Length - offset);
-            CurrentBufferLength = bytesRead + offset;
+            //The new buffer position is set to point at the byte that buffer
+            //  position pointed at (which is now at the head of the buffer)
             CurrentBufferPosition = 0;
 
-            return bytesRead != 0;
+            return true;
         }
 
         public bool Run()
         {
-            bool modified = false;
-            int lastWritten = CurrentBufferPosition;
-            int writtenSinceFlush = lastWritten;
+            int nextSequenceNumberThatCouldBeWritten = 0;
+            int bytesWrittenSinceLastFlush = 0;
+            bool anyOperationsExecuted = false;
 
-            if(CurrentBufferPosition == CurrentBufferLength)
+            while (true)
             {
-                AdvanceBuffer(CurrentBufferPosition);
-                lastWritten = 0;
-            }
-
-            while (CurrentBufferLength > 0)
-            {
-                int token;
-                int posedPosition = CurrentBufferPosition;
-
-                if (CurrentBufferLength == CurrentBuffer.Length && CurrentBufferLength == CurrentBufferPosition)
+                //Loop until we run out of data in the buffer
+                for(; CurrentBufferPosition < CurrentBufferLength;)
                 {
-                    int writeCount = CurrentBufferLength - lastWritten;
-
-                    if (writeCount > 0)
+                    int posedPosition = CurrentSequenceNumber;
+                    bool skipAdvanceBuffer = false;
+                    if (_trie.Accept(CurrentBuffer[CurrentBufferPosition], ref posedPosition, out TerminalLocation<OperationTerminal> terminal))
                     {
-                        _target.Write(CurrentBuffer, lastWritten, writeCount);
-                        writtenSinceFlush += writeCount;
+                        IOperation operation = terminal.Terminal.Operation;
+                        int matchLength = terminal.Terminal.End - terminal.Terminal.Start;
+                        int handoffBufferPosition = CurrentBufferPosition + matchLength - (CurrentSequenceNumber - terminal.Location);
+
+                        if (terminal.Location > nextSequenceNumberThatCouldBeWritten)
+                        {
+                            int toWrite = terminal.Location - nextSequenceNumberThatCouldBeWritten;
+                            _target.Write(CurrentBuffer, handoffBufferPosition - toWrite - matchLength, toWrite);
+                            bytesWrittenSinceLastFlush += toWrite;
+                            nextSequenceNumberThatCouldBeWritten = posedPosition - matchLength;
+                        }
+
+                        if(operation.Id == null || (Config.Flags.TryGetValue(operation.Id, out bool opEnabledFlag) && opEnabledFlag))
+                        {
+                            CurrentSequenceNumber += handoffBufferPosition - CurrentBufferPosition;
+                            CurrentBufferPosition = handoffBufferPosition;
+                            posedPosition = handoffBufferPosition;
+                            int bytesWritten = operation.HandleMatch(this, CurrentBufferLength, ref posedPosition, terminal.Terminal.Token, _target);
+                            bytesWrittenSinceLastFlush += bytesWritten;
+
+                            CurrentSequenceNumber += posedPosition - CurrentBufferPosition;
+                            CurrentBufferPosition = posedPosition;
+                            nextSequenceNumberThatCouldBeWritten = CurrentSequenceNumber;
+                            skipAdvanceBuffer = true;
+                            anyOperationsExecuted = true;
+                        }
+                        else
+                        {
+                            int oldSequenceNumber = CurrentSequenceNumber;
+                            CurrentSequenceNumber = terminal.Location + terminal.Terminal.End;
+                            CurrentBufferPosition += CurrentSequenceNumber - oldSequenceNumber;
+                        }
+
+                        if (bytesWrittenSinceLastFlush >= _flushThreshold)
+                        {
+                            _target.Flush();
+                        }
                     }
 
-                    AdvanceBuffer(CurrentBufferLength - (CurrentSequenceNumber - _trie.OldestRequiredSequenceNumber));
-                    lastWritten = 0;
-                    posedPosition = 0;
+                    if (!skipAdvanceBuffer)
+                    {
+                        ++CurrentSequenceNumber;
+                        ++CurrentBufferPosition;
+                    }
                 }
 
-                int sn = CurrentSequenceNumber;
-                bool isMatch = _trie.Accept(CurrentBuffer[CurrentBufferPosition], ref sn, out TerminalLocation<OperationTerminal> terminal);
-                IOperation op = isMatch ? terminal.Terminal.Operation : null;
-                sn = isMatch ? terminal.Location + terminal.Terminal.End - terminal.Terminal.Start : sn;
-                CurrentBufferPosition -= CurrentSequenceNumber - sn;
-                CurrentSequenceNumber = sn;
-                token = isMatch ? terminal.Terminal.Token : -1;
-                bool opEnabledFlag;
+                //Calculate the sequence number at the head of the buffer
+                int headSequenceNumber = CurrentSequenceNumber - CurrentBufferPosition;
+                int bufferPositionToAdvanceTo = _trie.OldestRequiredSequenceNumber - headSequenceNumber;
+                int numberOfUncommittedBytesBeforeThePositionToAdvanceTo = _trie.OldestRequiredSequenceNumber - nextSequenceNumberThatCouldBeWritten;
 
-                if ((op != null)
-                        && ((op.Id == null)
-                            || (Config.Flags.TryGetValue(op.Id, out opEnabledFlag) && opEnabledFlag))
-                    )
+                //If we'd advance data out of the buffer that hasn't been
+                //  handled already, write it out
+                if (numberOfUncommittedBytesBeforeThePositionToAdvanceTo > 0)
                 {
-                    // The operation will be processed because one of these conditions are met:
-                    // - The operation doesn't have an id (thus can't be disabled)
-                    // - The flag for the Id exists and is true.
-
-                    int writeCount = CurrentBufferPosition - (CurrentSequenceNumber - terminal.Location) - lastWritten;
-
-                    if (writeCount > 0)
-                    {
-                        _target.Write(CurrentBuffer, lastWritten, writeCount);
-                        writtenSinceFlush += writeCount;
-                    }
-
-                    //Advance the sequence number by the number of bytes taken by the token
-                    CurrentSequenceNumber += posedPosition - CurrentBufferPosition;
-                    CurrentBufferPosition = posedPosition;
-
-                    try
-                    {
-                        writtenSinceFlush += op.HandleMatch(this, CurrentBufferLength, ref posedPosition, token, _target);
-                        CurrentSequenceNumber += posedPosition - CurrentBufferPosition;
-                    }
-                    catch (Exception ex)
-                    {
-                        throw new Exception($"Error running handler {op} at position {CurrentBufferPosition} in {Encoding.EncodingName} bytes of {Encoding.GetString(CurrentBuffer, 0, CurrentBufferLength)}.\n\nStart: {Encoding.GetString(CurrentBuffer, CurrentBufferPosition, CurrentBufferLength - CurrentBufferPosition)} \n\nCheck InnerException for details.", ex);
-                    }
-
-                    CurrentBufferPosition = posedPosition;
-                    lastWritten = posedPosition;
-                    modified = true;
-                }
-                else
-                {
-                    // If the operation is disabled, it's as if the token were just arbitrary text, so do nothing special with it.
-                    ++CurrentBufferPosition;
-                    ++CurrentSequenceNumber;
+                    int toWrite = numberOfUncommittedBytesBeforeThePositionToAdvanceTo;
+                    _target.Write(CurrentBuffer, bufferPositionToAdvanceTo - toWrite, toWrite);
+                    bytesWrittenSinceLastFlush += toWrite;
+                    nextSequenceNumberThatCouldBeWritten = _trie.OldestRequiredSequenceNumber;
                 }
 
-                if (CurrentBufferPosition == CurrentBufferLength)
+                //We ran out of data in the buffer, so attempt to advance
+                //  if we fail, 
+                if (!AdvanceBuffer(bufferPositionToAdvanceTo))
                 {
-                    int writeCount = CurrentBufferPosition - lastWritten - (CurrentSequenceNumber - _trie.OldestRequiredSequenceNumber);
+                    int posedPosition = CurrentSequenceNumber;
+                    _trie.FinalizeMatchesInProgress(ref posedPosition, out TerminalLocation<OperationTerminal> terminal);
 
-                    if (writeCount > 0)
+                    while (terminal != null)
                     {
-                        _target.Write(CurrentBuffer, lastWritten, writeCount);
-                        writtenSinceFlush += writeCount;
+                        IOperation operation = terminal.Terminal.Operation;
+                        int matchLength = terminal.Terminal.End - terminal.Terminal.Start;
+                        int handoffBufferPosition = CurrentBufferPosition + matchLength - (CurrentSequenceNumber - terminal.Location);
+
+                        if (terminal.Location > nextSequenceNumberThatCouldBeWritten)
+                        {
+                            int toWrite = terminal.Location - nextSequenceNumberThatCouldBeWritten;
+                            _target.Write(CurrentBuffer, handoffBufferPosition - toWrite - matchLength, toWrite);
+                            bytesWrittenSinceLastFlush += toWrite;
+                            nextSequenceNumberThatCouldBeWritten = terminal.Location;
+                        }
+
+                        if (operation.Id == null || (Config.Flags.TryGetValue(operation.Id, out bool opEnabledFlag) && opEnabledFlag))
+                        {
+                            CurrentSequenceNumber += handoffBufferPosition - CurrentBufferPosition;
+                            CurrentBufferPosition = handoffBufferPosition;
+                            posedPosition = handoffBufferPosition;
+                            int bytesWritten = operation.HandleMatch(this, CurrentBufferLength, ref posedPosition, terminal.Terminal.Token, _target);
+                            bytesWrittenSinceLastFlush += bytesWritten;
+
+                            CurrentSequenceNumber += posedPosition - CurrentBufferPosition;
+                            CurrentBufferPosition = posedPosition;
+                            nextSequenceNumberThatCouldBeWritten = CurrentSequenceNumber;
+                            anyOperationsExecuted = true;
+                        }
+                        else
+                        {
+                            int oldSequenceNumber = CurrentSequenceNumber;
+                            CurrentSequenceNumber = terminal.Location + terminal.Terminal.End;
+                            CurrentBufferPosition += CurrentSequenceNumber - oldSequenceNumber;
+                        }
+
+                        _trie.FinalizeMatchesInProgress(ref posedPosition, out terminal);
                     }
 
-                    int advanceBufferMax = CurrentBufferLength - (CurrentSequenceNumber - _trie.OldestRequiredSequenceNumber);
-                    if (!AdvanceBuffer(Math.Min(CurrentBufferPosition, Math.Max(0, advanceBufferMax))) && CurrentBufferPosition == CurrentBufferLength)
-                    {
-                        break;
-                    }
-
-                    lastWritten = 0;
-                }
-
-                if (writtenSinceFlush >= _flushThreshold)
-                {
-                    writtenSinceFlush = 0;
-                    _target.Flush();
-                }
-            }
-
-            int n = CurrentSequenceNumber;
-            _trie.FinalizeMatchesInProgress(ref n, out TerminalLocation<OperationTerminal> term);
-
-            if(term != null)
-            {
-                IOperation op = term.Terminal.Operation;
-                int sn = term.Location + term.Terminal.End - term.Terminal.Start;
-                CurrentBufferPosition -= CurrentSequenceNumber - sn;
-                CurrentSequenceNumber = sn;
-                bool opEnabledFlag;
-
-                if ((op != null)
-                        && ((op.Id == null)
-                            || (Config.Flags.TryGetValue(op.Id, out opEnabledFlag) && opEnabledFlag))
-                    )
-                {
-                    int pos = CurrentBufferPosition;
-                    op.HandleMatch(this, CurrentBufferLength, ref pos, term.Terminal.Token, _target);
-                    lastWritten = pos;
-                    _target.Flush();
-                    modified = true;
+                    break;
                 }
             }
 
-            if (lastWritten < CurrentBufferPosition)
+            int endSequenceNumber = CurrentSequenceNumber - CurrentBufferPosition + CurrentBufferLength;
+            if (endSequenceNumber > nextSequenceNumberThatCouldBeWritten)
             {
-                int writeCount = CurrentBufferPosition - lastWritten;
-
-                if (writeCount > 0)
-                {
-                    _target.Write(CurrentBuffer, lastWritten, writeCount);
-                }
+                int toWrite = endSequenceNumber - nextSequenceNumberThatCouldBeWritten;
+                _target.Write(CurrentBuffer, CurrentBufferLength - toWrite, toWrite);
             }
 
             _target.Flush();
-            return modified;
+            return anyOperationsExecuted;
         }
 
         public void SeekBackUntil(ITokenTrie match)
@@ -403,14 +401,39 @@ namespace Microsoft.TemplateEngine.Core.Util
             }
         }
 
-        public void SeekForwardThrough(ITokenTrie match, ref int bufferLength, ref int currentBufferPosition)
+        public void SeekForwardThrough(ITokenTrie trie, ref int bufferLength, ref int currentBufferPosition)
         {
-            BaseSeekForward(match, ref bufferLength, ref currentBufferPosition, true);
+            BaseSeekForward(trie, ref bufferLength, ref currentBufferPosition, true);
         }
 
-        public void SeekForwardUntil(ITokenTrie match, ref int bufferLength, ref int currentBufferPosition)
+        public void SeekForwardUntil(ITokenTrie trie, ref int bufferLength, ref int currentBufferPosition)
         {
-            BaseSeekForward(match, ref bufferLength, ref currentBufferPosition, false);
+            BaseSeekForward(trie, ref bufferLength, ref currentBufferPosition, false);
+        }
+
+        public void SeekForwardWhile(ITokenTrie trie, ref int bufferLength, ref int currentBufferPosition)
+        {
+            while (bufferLength > trie.MinLength)
+            {
+                while (currentBufferPosition < bufferLength - trie.MinLength + 1)
+                {
+                    if (bufferLength == 0)
+                    {
+                        currentBufferPosition = 0;
+                        return;
+                    }
+
+                    int token;
+                    if (!trie.GetOperation(CurrentBuffer, bufferLength, ref currentBufferPosition, out token))
+                    {
+                        return;
+                    }
+                }
+
+                AdvanceBuffer(currentBufferPosition);
+                currentBufferPosition = CurrentBufferPosition;
+                bufferLength = CurrentBufferLength;
+            }
         }
 
         private void BaseSeekForward(ITokenTrie match, ref int bufferLength, ref int currentBufferPosition, bool consumeToken)
@@ -450,31 +473,6 @@ namespace Microsoft.TemplateEngine.Core.Util
 
             //Ran out of places to check and haven't reached the actual match, consume all the way to the end
             currentBufferPosition = bufferLength;
-        }
-
-        public void SeekForwardWhile(ITokenTrie match, ref int bufferLength, ref int currentBufferPosition)
-        {
-            while (bufferLength > match.MinLength)
-            {
-                while (currentBufferPosition < bufferLength - match.MinLength + 1)
-                {
-                    if (bufferLength == 0)
-                    {
-                        currentBufferPosition = 0;
-                        return;
-                    }
-
-                    int token;
-                    if (!match.GetOperation(CurrentBuffer, bufferLength, ref currentBufferPosition, out token))
-                    {
-                        return;
-                    }
-                }
-
-                AdvanceBuffer(currentBufferPosition);
-                currentBufferPosition = CurrentBufferPosition;
-                bufferLength = CurrentBufferLength;
-            }
         }
     }
 }
