+++ src/Microsoft.TemplateEngine.Orchestrator.RunnableProjects/Config/ReplacementConfig.cs
@@ -21,7 +21,7 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects.Config
             if (parameters.TryGetRuntimeValue(environmentSettings, tokens.VariableName, out object newValueObject))
             {
                 string newValue = newValueObject.ToString();
-                return new Replacement(tokens.OriginalValue.TokenConfig(), newValue, null);
+                return new Replacement(tokens.OriginalValue, newValue, null);
             }
             else
             {
@@ -36,7 +36,39 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects.Config
             string replacement = rawConfiguration.ToString("replacement");
             string id = rawConfiguration.ToString("id");
 
-            yield return new Replacement(original.TokenConfig(), replacement, id);
+            JArray onlyIf = rawConfiguration.Get<JArray>("onlyIf");
+            TokenConfig coreConfig = original.TokenConfigBuilder();
+
+            if (onlyIf != null)
+            {
+                foreach (JToken entry in onlyIf.Children())
+                {
+                    if (!(entry is JObject x))
+                    {
+                        continue;
+                    }
+
+                    string before = entry.ToString("before");
+                    string after = entry.ToString("after");
+                    TokenConfig entryConfig = coreConfig;
+
+                    if (!string.IsNullOrEmpty(before))
+                    {
+                        entryConfig = entryConfig.OnlyIfBefore(before);
+                    }
+
+                    if (!string.IsNullOrEmpty(after))
+                    {
+                        entryConfig = entryConfig.OnlyIfAfter(after);
+                    }
+
+                    yield return new Replacement(entryConfig, replacement, id);
+                }
+            }
+            else
+            {
+                yield return new Replacement(coreConfig, replacement, id);
+            }
         }
     }
 }
