+++ src/Microsoft.TemplateEngine.Orchestrator.RunnableProjects/GeneratedSymbol.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Microsoft.TemplateEngine.Utils;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
@@ -19,6 +20,8 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
 
         public string Type { get; set; }
 
+        public IReadOnlyList<IReplacementContext> ReplacementContexts { get; set; }
+
         public static ISymbolModel FromJObject(JObject jObject)
         {
             GeneratedSymbol sym = new GeneratedSymbol
@@ -27,10 +30,38 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
                 Generator = jObject.ToString(nameof(Generator)),
                 Parameters = jObject.ToJTokenDictionary(StringComparer.Ordinal, nameof(Parameters)),
                 Type = jObject.ToString(nameof(Type)),
-                Replaces = jObject.ToString(nameof(Replaces))
+                Replaces = jObject.ToString(nameof(Replaces)),
+                ReplacementContexts = ReadReplacementContexts(jObject)
             };
 
             return sym;
+        }
+
+        private static IReadOnlyList<IReplacementContext> ReadReplacementContexts(JObject jObject)
+        {
+            JArray onlyIf = jObject.Get<JArray>("onlyIf");
+
+            if (onlyIf != null)
+            {
+                List<IReplacementContext> contexts = new List<IReplacementContext>();
+                foreach (JToken entry in onlyIf.Children())
+                {
+                    if (!(entry is JObject x))
+                    {
+                        continue;
+                    }
+
+                    string before = entry.ToString("before");
+                    string after = entry.ToString("after");
+                    contexts.Add(new ReplacementContext(before, after));
+                }
+
+                return contexts;
+            }
+            else
+            {
+                return Empty<IReplacementContext>.List.Value;
+            }
         }
     }
 }
