+++ src/Microsoft.TemplateEngine.Orchestrator.RunnableProjects/ISymbolModel.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
 {
     public interface ISymbolModel
@@ -7,5 +9,14 @@ namespace Microsoft.TemplateEngine.Orchestrator.RunnableProjects
         string Binding { get; set; }
 
         string Replaces { get; set; }
+
+        IReadOnlyList<IReplacementContext> ReplacementContexts { get; }
+    }
+
+    public interface IReplacementContext
+    {
+        string OnlyIfBefore { get; }
+
+        string OnlyIfAfter { get; }
     }
 }
